+++ src/LaravelDddDomainResourcesServiceProvider.php
@@ -4,9 +4,10 @@ declare(strict_types=1);
 
 namespace Alibori\LaravelDddDomainResources;
 
+use Alibori\LaravelDddDomainResources\Commands\Generators\LaravelDddDomainResourcesGeneratorCommand;
+use Alibori\LaravelDddDomainResources\Commands\LaravelDddDomainResourcesCommand;
 use Spatie\LaravelPackageTools\Package;
 use Spatie\LaravelPackageTools\PackageServiceProvider;
-use Alibori\LaravelDddDomainResources\Commands\LaravelDddDomainResourcesCommand;
 
 class LaravelDddDomainResourcesServiceProvider extends PackageServiceProvider
 {
@@ -15,6 +16,17 @@ class LaravelDddDomainResourcesServiceProvider extends PackageServiceProvider
         $package
             ->name(name: 'laravel-ddd-domain-resources')
             ->hasConfigFile(configFileName: 'ddd-domain-resources')
-            ->hasCommand(commandClassName: LaravelDddDomainResourcesCommand::class);
+            ->hasCommand(commandClassName: LaravelDddDomainResourcesCommand::class)
+            ->hasCommands(commandClassNames: LaravelDddDomainResourcesGeneratorCommand::class);
+    }
+
+    public function packageRegistered(): void
+    {
+        $this->publishes(
+            paths:[
+                __DIR__.'/../resources/stubs/' => base_path(path: 'stubs/ddd-domain-resources'),
+            ],
+            groups: 'ddd-domain-resources-stubs'
+        );
     }
 }