+++ CHANGELOG.md
@@ -1,3 +1,19 @@
 # Changelog
 
 All notable changes to `laravel-ddd-domain-resources` will be documented in this file.
+
+## 1.0.0 - 2023-04-XX
+
+### Added
+
+Features:
+
+-  Added flag to generate a domain directory structure.
+-  Added option to generate the following domain resources:
+    -  Use Cases
+    -  Contracts
+    -  Events
+    -  Exceptions
+    -  Value Objects
+    -  Repositories
+    -  Controllers